Updated to allow creating and deleting users
In addtion we are moving some of the database management into a more
generic location so that when I do the script management I can utilize
the same functions. This will let me expand further in the future
however I'd like
